// apps/admin/src/generated/api/schemas/listAuditLogsParams.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Common Admin API
 * API for the common admin starter template
 * OpenAPI spec version: 0.1.0
 */
import type { ListAuditLogsAction } from "./listAuditLogsAction";
import type { ListAuditLogsResourceType } from "./listAuditLogsResourceType";
import type { Object } from "./object";

export type ListAuditLogsParams = {
  /**
   * @minimum 1
   */
  page?: Object;
  /**
   * @minimum 1
   * @maximum 100
   */
  pageSize?: Object;
  search?: string;
  /**
   * Sort by one of: createdAt, action, resourceType, actorEmail, actorName
   */
  sort?: string;
  actorUserId?: string;
  action?: ListAuditLogsAction;
  resourceType?: ListAuditLogsResourceType;
  resourceId?: string;
  dateFrom?: string;
  dateTo?: string;
};